separated logic into hook
